Move change types into their own file
